Merged the HDF5 and FS interfaces into a single object.
This makes it easier to open existing HDF5 files or to create a new HDF5 file,
as the jaspagate code doesn't have to pass information between two interfaces.
This is particularly relevant when a local filesystem does not exist and we
can't just rely on passing around paths to communicate between interfaces.